t1_loader/requirements.txt
kafka-python
psycopg2
SQLAlchemy==1.2.15
